% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{Individual}
\alias{Individual}
\title{Individual R6 Class}
\description{
Individual R6 Class

Individual R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Individual-new}{\code{Individual$new()}}
\item \href{#method-Individual-print}{\code{Individual$print()}}
\item \href{#method-Individual-calculate_bmi}{\code{Individual$calculate_bmi()}}
\item \href{#method-Individual-validate_sex}{\code{Individual$validate_sex()}}
\item \href{#method-Individual-validate_height_value}{\code{Individual$validate_height_value()}}
\item \href{#method-Individual-validate_weight_value}{\code{Individual$validate_weight_value()}}
\item \href{#method-Individual-validate_age_value}{\code{Individual$validate_age_value()}}
\item \href{#method-Individual-validate_numeric}{\code{Individual$validate_numeric()}}
\item \href{#method-Individual-get_age_value}{\code{Individual$get_age_value()}}
\item \href{#method-Individual-print_individual}{\code{Individual$print_individual()}}
\item \href{#method-Individual-get_individual_df}{\code{Individual$get_individual_df()}}
\item \href{#method-Individual-clone}{\code{Individual$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Individual-new"></a>}}
\if{latex}{\out{\hypertarget{method-Individual-new}{}}}
\subsection{Method \code{new()}}{
Initialize the individual's registry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$new(sex, age, height, weight)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sex}}{character value has two options: Male, Female (Case doesn't matter)}

\item{\code{age}}{numeric value. (range: \link{1 - 110})}

\item{\code{height}}{numeric value. (range: \link{0.4 - 2.2}) (meters)}

\item{\code{weight}}{numeric value. (range: \link{1 - 250}) (kg)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Individual-print"></a>}}
\if{latex}{\out{\hypertarget{method-Individual-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Individual-calculate_bmi"></a>}}
\if{latex}{\out{\hypertarget{method-Individual-calculate_bmi}{}}}
\subsection{Method \code{calculate_bmi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$calculate_bmi(weight, height)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Individual-validate_sex"></a>}}
\if{latex}{\out{\hypertarget{method-Individual-validate_sex}{}}}
\subsection{Method \code{validate_sex()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$validate_sex(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Individual-validate_height_value"></a>}}
\if{latex}{\out{\hypertarget{method-Individual-validate_height_value}{}}}
\subsection{Method \code{validate_height_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$validate_height_value(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Individual-validate_weight_value"></a>}}
\if{latex}{\out{\hypertarget{method-Individual-validate_weight_value}{}}}
\subsection{Method \code{validate_weight_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$validate_weight_value(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Individual-validate_age_value"></a>}}
\if{latex}{\out{\hypertarget{method-Individual-validate_age_value}{}}}
\subsection{Method \code{validate_age_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$validate_age_value(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Individual-validate_numeric"></a>}}
\if{latex}{\out{\hypertarget{method-Individual-validate_numeric}{}}}
\subsection{Method \code{validate_numeric()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$validate_numeric(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Individual-get_age_value"></a>}}
\if{latex}{\out{\hypertarget{method-Individual-get_age_value}{}}}
\subsection{Method \code{get_age_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$get_age_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Individual-print_individual"></a>}}
\if{latex}{\out{\hypertarget{method-Individual-print_individual}{}}}
\subsection{Method \code{print_individual()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$print_individual()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Individual-get_individual_df"></a>}}
\if{latex}{\out{\hypertarget{method-Individual-get_individual_df}{}}}
\subsection{Method \code{get_individual_df()}}{
Get individual in df format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$get_individual_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Individual-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Individual-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Individual$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
